feat: contract call offchain data (#3)
* feat: rest-service skeleton

The REST API will be used by the users to provide the payload for the
contract call.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* feat(rest-service): call contract with offchain data

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: ignore `storage` file

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: remove dead code

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: update solana-axelar

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(rest-service): serialize test runs

Otherwise we try to bind the same port several times.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(rest-service): make shutdown sender available

We might need to send the shutdown signal from outside the service, thus
it's a good idea to make it available. The first place where we see the
need is in the tests as we need to gracefully shutdown the server after
every test to avoid trying to bind on the same port over and over.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix: rest-service binding error in tests

Seems like using the test_serial crate is not enough, nor shutting down
the http server gracefully. Let's try to limit the threads used by
nextest.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* feat: add health endpoint to signal the server is up

And use it in the tests instead of `sleep`.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>