feat: fee evaluation before task execution (#78)
* feat: fee estimation before execution

This patch adds a gas estimation module which is supposed to work with a
local node with a fork of the actual Solana state. The reason this is
necessary is because due to transaction size limitations on Solana and
Axelar minimum requirements, we need to perform Axelar tasks through
multiple Solana transactions. Before performing the task though, we need
to know if the available gas is enough to cover the costs of all
transactions required for that: initializing the message payload,
uploading the payload on-chain, committing the payload, invoking the
destination program with the payload account, and then closing the
payload account.

Normally the estimation is done through the `simulate_transaction` call
on the Solana RpcClient, but since multiple transactions are required
for an Axelar task, the transactions that need accounts created by
earlier transactions would fail, since the transactions that create
them are not really broadcast on the real chain and these accounts
are thus not created on-chain.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix: return specific error so it can be handled

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(gateway-task-processor): cache gas estimations

When a task fails due to gas not being enough, the user can add gas and
the relayer receives the execute task again. If we don't cache and try
to estimate on the local validator a second time, the estimation fails.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor(gateway-task-processor): cache cu usage only

Instead of caching the cost only or not caching at all, we can cache the
resources used and recalculate the fees only.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* feat(gateway-task-processor): surfpool account cleanup

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(gateway-task-processor): several small fixes

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(gateway-task-processor): don't clean payer

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(gateway-task-processor): remove account cleanup

Setting the lamports to 0 doesn't seem to remove the account. Following
transactions using the account fail since it has 0 lamports and the data
is not fetched from the blockchain again.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: try remove ignore from test

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: update README

Instructions on how to use/setup surfpool

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>